requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
anthropic==0.39.0
pyyaml==6.0.1
requests==2.31.0